# Alerting: Improve performance of TrimResults